Improving data holding in point grids

+++ cg/include/geometry/PointHolder.h
@@ -1,6 +1,6 @@
 //[]---------------------------------------------------------------[]
 //|                                                                 |
-//| Copyright (C) 2019 Orthrus Group.                               |
+//| Copyright (C) 2019, 2021 Orthrus Group.                         |
 //|                                                                 |
 //| This software is provided 'as-is', without any express or       |
 //| implied warranty. In no event will the authors be held liable   |
@@ -28,7 +28,7 @@
 // Class definition for point holder.
 //
 // Author: Paulo Pagliosa
-// Last revision: 16/03/2019
+// Last revision: 08/12/2021
 
 #ifndef __PointHolder_h
 #define __PointHolder_h
@@ -61,7 +61,7 @@ protected:
     // do nothing
   }
 
-  template <size_t D, typename real>
+  template <int D, typename real>
   static auto computeBounds(const PointArray& points, bool squared)
   {
     using size_type = decltype(points.size());
